Exclusive new feature! Creates a multiboot ramdisk.
This ramdisk test if there is a multiboot_ramdisk.tar file on the external sdcard.
If this archive exist, it's extracted at rootfs /
Then if a script named /pre-init.sh exist, it's executed.

After that the (hopefully new) init file is excecuted

Note that the permissions on the mounted external sd are not yet set correctly